Add files via upload
The selection of the number of operating modes (clusters) was carried out using the "Elbow method". The essence of the method is that it is necessary to select such a number of clusters so that adding another one does not improve the quality of clustering.
i The criterion by which the quality of clustering was assessed with a different number of clusters, inertia (the sum of squared errors in the context of clustering)
The “fracture” characteristic of this method occurs on 2 clusters. Thus, 2 classes of states of the ACS GTE are distinguished. In a similar way, the number of state classes for other mobile objects can 
To make sure the data breaks down really well into the two state classes of a mobile object, visualization of this breakdown is needed. Visualization was carried out by two methods: principal component analysis (PCA)  and t-distributed Stochastic Neighbor Embedding (t-SNE) [10].
it is possible to determine the physical meaning of such a distribution of points. The increments of parameters close to zero indicate the static operating mode of the ACS GTE, the increments located far from the center of coordinates indicate the dynamic operating mode of the object.